Convert console functions tests to mocha
