+++ test/test-async.js
@@ -4,71 +4,3 @@
  */
 
 require('babel-core/register');
-var async = require('../lib');
-
-// generates tests for console functions such as async.log
-var console_fn_tests = function(name){
-
-    if (typeof console !== 'undefined') {
-        exports[name] = function(test){
-            test.expect(5);
-            var fn = function(arg1, callback){
-                test.equals(arg1, 'one');
-                setTimeout(function(){callback(null, 'test');}, 0);
-            };
-            var fn_err = function(arg1, callback){
-                test.equals(arg1, 'one');
-                setTimeout(function(){callback('error');}, 0);
-            };
-            var _console_fn = console[name];
-            var _error = console.error;
-            console[name] = function(val){
-                test.equals(val, 'test');
-                test.equals(arguments.length, 1);
-                console.error = function(val){
-                    test.equals(val, 'error');
-                    console[name] = _console_fn;
-                    console.error = _error;
-                    test.done();
-                };
-                async[name](fn_err, 'one');
-            };
-            async[name](fn, 'one');
-        };
-
-        exports[name + ' with multiple result params'] = function(test){
-            test.expect(1);
-            var fn = function(callback){callback(null,'one','two','three');};
-            var _console_fn = console[name];
-            var called_with = [];
-            console[name] = function(x){
-                called_with.push(x);
-            };
-            async[name](fn);
-            test.same(called_with, ['one','two','three']);
-            console[name] = _console_fn;
-            test.done();
-        };
-    }
-
-    // browser-only test
-    exports[name + ' without console.' + name] = function(test){
-        if (typeof window !== 'undefined') {
-            var _console = window.console;
-            window.console = undefined;
-            var fn = function(callback){callback(null, 'val');};
-            var fn_err = function(callback){callback('error');};
-            async[name](fn);
-            async[name](fn_err);
-            window.console = _console;
-        }
-        test.done();
-    };
-
-};
-
-console_fn_tests('log');
-console_fn_tests('dir');
-/*console_fn_tests('info');
-console_fn_tests('warn');
-console_fn_tests('error');*/
